move the very generic stuff off to the side
